+++ code/sys.ui/react.dev/src/Dev.ts
@@ -1,7 +1,7 @@
-import { DevBus as Bus } from './fn.Bus';
-import { Context } from './fn.Ctx';
-import { Spec } from './fn.Spec';
-import { ValueHandler } from './fn.Tools';
+import { DevBus as Bus } from './f.Bus';
+import { Context } from './f.Ctx';
+import { Spec } from './f.Spec';
+import { ValueHandler } from './f.Tools';
 import { headless } from './test.ui/headless';
 import { Harness } from './ui/Harness';
 import { ModuleList } from './ui/List.Module';
