+++ code/sys.ui/react.dev/src/f.Bus/Bus.Controller.mts
@@ -1,4 +1,4 @@
-import { Context } from '../fn.Ctx';
+import { Context } from '../f.Ctx';
 import { BusEvents } from './Bus.Events.mjs';
 import { BusMemoryState } from './Bus.MemoryState.mjs';
 import { DEFAULTS, Id, Is, R, Test, rx, type t } from './common';
