+++ code/sys.ui/react.dev/src/f.Bus/Bus.Events.mts
@@ -1,4 +1,4 @@
-import { ContextState } from '../fn.Ctx/Context.State.mjs';
+import { ContextState } from '../f.Ctx/Context.State.mjs';
 import { asArray, rx, slug, type t } from './common';
 
 const DEFAULT = { TIMEOUT: 500 };
