+++ code/sys.ui/react.dev/src/f.Bus/Bus.MemoryState.mts
