+++ code/sys.ui/react.dev/src/f.Bus/common.ts
@@ -1,5 +1,2 @@
 export * from '../common';
-export * from '../f.Bus';
-export * from '../ui.use';
-
 export { Spec } from '../f.Spec';
