+++ code/sys.ui/react.dev/src/f.Ctx/Context.mts
@@ -1,6 +1,6 @@
 import { type t } from './common';
 
-import { BusEvents } from '../fn.Bus/Bus.Events.mjs';
+import { BusEvents } from '../f.Bus/Bus.Events.mjs';
 import { ContextState } from './Context.State.mjs';
 import { CtxProps } from './Ctx.Props.mjs';
