+++ code/sys.ui/react.dev/src/f.Ctx/Ctx.PanelEdge.mts
