+++ code/sys.ui/react.dev/src/f.Ctx/common.types.ts
