--- code/sys.ui/react.dev/src/fn.Bus/common.ts
@@ -1,2 +1,0 @@
-export * from '../common';
-export { Spec } from '../fn.Spec';
