+++ code/sys.ui/react.dev/src/test.ui/index.ts
@@ -1,4 +1,4 @@
-export { Spec } from '../fn.Spec';
+export { Spec } from '../f.Spec';
 export { Dev } from '../Dev';
 
 export * from './common';
