+++ code/sys.ui/react.dev/src/ui.use/common.ts
@@ -1,2 +1,2 @@
 export * from '../common';
-export { DevBus } from '../fn.Bus';
+export { DevBus } from '../f.Bus';
