+++ code/sys.ui/react.dev/src/ui.use/useRenderer.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import { DevBus } from '../fn.Bus';
+import { DevBus } from '../f.Bus';
 import { DEFAULTS, Is, type t } from './common';
 import { useCurrentState } from './useCurrentState';
 import { useRedrawEvent } from './useRedrawEvent';
